requirements.txt
omniverse==0.0.36
peft==0.11.1
transformers==4.41.2
modal==0.62.201
pandas-stubs==2.2.2.240603
rich==13.7.1